pyproject.toml
[project]
name = "powerclip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
    "pyperclip>=1.11.0",
]
